+++ NEWS.md
@@ -1,3 +1,36 @@
+# Changes in mirt 1.44
+
+- Added `Attitude` dataset from Andrich (1988) publication to demonstrate
+  constrained hyperbolic cosine model (HCM) and the same model with the 
+  estimated latitude of acceptance parameters
+
+- Added several `itemtype` inputs to specify a family of unfolding models 
+  for dichotomous and polytomous data. These currently include unidimensional 
+  and multidimensional versions of the 
+  (generalized) hyperbolic cosine model, (generalized) absolute logistic model,
+  (generalized) simple squared logistic model, and the 
+  (generalized) parallellogram analysis model. `simdata()` also gained 
+  support for the Luo (2001) family of generating models, among most other 
+  secondary functions
+
+- `summary()` now automatically outputs delta-method SEs for 
+  standardized factor loadings.
+  Only applied for non-EFA models that include an estimate of the ACOV 
+  (e.g., via `mirt(..., SE=TRUE)`)
+
+- Standardized factor loadings for the multidimensional nominal response model
+  now report consistent values regardless of the category ordering
+
+- Exported general-purpose `DeltaMethod()` function for numerical 
+  version of the delta method
+  
+- Fixed `M2()` computations with large amounts of missing data, particularly
+  prevalent with the C2 statistic (reported by Hynek Cigler)
+  
+- Fixed minor `coef(..., IRTpars=TRUE)` issue where output reported a 
+  constant SE term of 0.0 when converting Rasch 
+  models (bk = -dk were not correctly tracked)
+
 # Changes in mirt 1.43
 
 - `M2()` family no longer requires row-wise removal of missing data to behave